+++ src/tokens/base.ts
@@ -95,6 +95,8 @@ export enum CoinKey {
   USDm = 'USDm', // Stablecoin on MegaETH: MegaUSD
   USD1 = 'USD1', // World Liberty Financial USD
   PathUSD = 'PathUSD', // Stablecoin on Tempo chain
+  frxUSD = 'frxUSD', // Stablecoin on Fraxtal chain
+  EURC = 'EURC', // Stablecoin for Euro
 
   // Other tokens
   WBTC = 'WBTC',